+++ codex-rs/core/src/client/mod.rs
@@ -1,0 +1,18 @@
+mod chat_completions;
+pub mod http;
+mod responses;
+pub mod types;
+
+pub(crate) use chat_completions::AggregateStreamExt;
+pub(crate) use chat_completions::AggregatedChatStream;
+pub(crate) use chat_completions::stream_chat_completions;
+pub use responses::ModelClient;
+pub(crate) use types::FreeformTool;
+pub(crate) use types::FreeformToolFormat;
+pub(crate) use types::Reasoning;
+pub use types::ResponseEvent;
+pub use types::ResponseStream;
+pub(crate) use types::ResponsesApiRequest;
+pub(crate) use types::ResponsesApiTool;
+pub(crate) use types::ToolSpec;
+pub(crate) use types::create_text_param_for_request;
