+++ codex-rs/core/src/tools/handlers/plan.rs
@@ -1,5 +1,5 @@
-use crate::client_common::tools::ResponsesApiTool;
-use crate::client_common::tools::ToolSpec;
+use crate::client::ResponsesApiTool;
+use crate::client::ToolSpec;
 use crate::codex::Session;
 use crate::codex::TurnContext;
 use crate::function_tool::FunctionCallError;
